Assert DOM equality across POI versions instead of skipping
POI 5.2.3 introduced bug-51037 in XSSFRow.onDocumentWrite, which
emits the default cell-style attribute (s="0") on every unstyled
cell. POI 4.x ~ 5.2.2 omit it. The SSML writer always emits it,
so a byte-equal comparison against POI's output diverges on the
older POI versions.

DOM-equivalence tests used to side-step that divergence by skipping
the whole assertion on POI <5.2.3 via Assumptions.assumeTrue, which
turned the compat-min CI job into a no-op for these cases
(skipped=6 in SSMLSheetWriterDomEquivalenceTest, skipped=9 in
RealisticAccuracyTest).

Extract the DOM-assertion helpers into XlsxDomAssertions and have
the IgnoringDimension variant strip the default s="0" on the SSML
side only when the POI side is the older one. Drop the now-dead
assumeTrue guards. Result: 15 cases that previously skipped on POI
4.1.1 now assert across both matrix entries, while the SSMLSheetReader
strict-format case keeps its assumeTrue (genuine POI 5.1.0+ capability
gate, not just an output divergence).
